+++ src/components/ids-data-grid/demos/columns-groups.ts
@@ -1,16 +1,17 @@
 import type IdsDataGrid from '../ids-data-grid';
 import '../ids-data-grid';
 import type { IdsDataGridColumn } from '../ids-data-grid-column';
+import IdsGlobal from '../../ids-global/ids-global';
 import booksJSON from '../../../assets/data/books.json';
 
 // Example for populating the DataGrid
 const dataGrid = document.querySelector<IdsDataGrid>('#data-grid-column-groups');
-const container: any = document.querySelector('ids-container');
+const locale = IdsGlobal.getLocale();
 
 if (dataGrid) {
   (async function init() {
     // Set Locale and wait for it to load
-    await container?.setLocale('en-US');
+    await locale.setLocale('en-US');
 
     // Do an ajax request
     const url: any = booksJSON;