+++ src/components/ids-data-grid/demos/contextmenu-thru-data.ts
@@ -2,15 +2,16 @@ import '../ids-data-grid';
 import type { IdsMenuObjectData, IdsMenuGroupData, IdsMenuItemData } from '../../ids-menu/ids-menu-attributes';
 
 import booksJSON from '../../../assets/data/books.json';
+import IdsGlobal from '../../ids-global/ids-global';
 
 // Example for populating the DataGrid
 const dataGrid: any = document.querySelector('#data-grid-contextmenu-thru-data');
-const container: any = document.querySelector('ids-container');
+const locale = IdsGlobal.getLocale();
 
 if (dataGrid) {
   (async function init() {
     // Set Locale and wait for it to load
-    await container?.setLocale('en-US');
+    await locale.setLocale('en-US');
 
     // Do an ajax request
     const url: any = booksJSON;