+++ src/components/ids-locale/ids-locale-number-options.ts
@@ -1,0 +1,8 @@
+export type IdsLocaleNumberOptions = Intl.NumberFormatOptions & {
+  locale?: string,
+  style?: string,
+  minimumFractionDigits?: number,
+  maximumFractionDigits?: number,
+  group?: string,
+  useGrouping?: boolean
+};