+++ src/components/ids-lookup/demos/pagination-client.ts
@@ -2,14 +2,14 @@
 import '../ids-lookup';
 
 import booksJSON from '../../../assets/data/books.json';
+import IdsGlobal from '../../ids-global/ids-global';
 
 // Example for populating the DataGrid
 const lookup: Element | any = document.querySelector('#lookup-1');
-const container: Element | any = document.querySelector('ids-container');
 
 (async function init() {
   // Set a Locale and wait for it to load
-  await container.setLocale('en-US');
+  await IdsGlobal.getLocale().setLocale('en-US');
 
   // Do an ajax request
   const url: any = booksJSON;